[refactoring] step4: 딜러 손익 계산 위치 수정
