+++ src/test/kotlin/blackjack_dealer/ParticipantsTest.kt
@@ -26,208 +26,4 @@ class ParticipantsTest : StringSpec({
             cardDeque = { listOf(Card(CardNumber.TWO, CardShape.CLOVER)).toGamerCards() },
         ).size shouldBe expected
     }
-
-    "참가자가 승,승 인경우 딜러의 손실은 두 참가자의 금액을 합한 값의 음수 이다" {
-        val input = "pita,haero"
-        val expected = -3000
-        var firstInput = true
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000), ParticipantBetAmount(2000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                if (firstInput) {
-                    firstInput = false
-                    listOf(
-                        Card(CardNumber.Q, CardShape.CLOVER),
-                        Card(CardNumber.J, CardShape.CLOVER),
-                    ).toGamerCards()
-                } else {
-                    listOf(
-                        Card(CardNumber.Q, CardShape.HEART),
-                        Card(CardNumber.THREE, CardShape.HEART),
-                    ).toGamerCards()
-                }
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.TWO, CardShape.CLOVER),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 20
-        participants.last().getCurrentCards().getCurrentScore() shouldBe 13
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
-
-    "참가자가 승,패 인경우 딜러의 손실은 두 참가자의 금액을 합한 값의 음수 이다" {
-        val input = "pita,haero"
-        val expected = 1000
-        var firstInput = true
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000), ParticipantBetAmount(2000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                if (firstInput) {
-                    firstInput = false
-                    listOf(
-                        Card(CardNumber.Q, CardShape.CLOVER),
-                        Card(CardNumber.J, CardShape.CLOVER),
-                    ).toGamerCards()
-                } else {
-                    listOf(
-                        Card(CardNumber.Q, CardShape.HEART),
-                        Card(CardNumber.THREE, CardShape.HEART),
-                    ).toGamerCards()
-                }
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.FIVE, CardShape.CLOVER),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 20
-        participants.last().getCurrentCards().getCurrentScore() shouldBe 13
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
-
-    "참가자가 패,패 인경우 딜러의 손실은 두 참가자의 금액을 합한 값의 음수 이다" {
-        val input = "pita,haero"
-        val expected = 3000
-        var firstInput = true
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000), ParticipantBetAmount(2000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                if (firstInput) {
-                    firstInput = false
-                    listOf(
-                        Card(CardNumber.Q, CardShape.CLOVER),
-                        Card(CardNumber.FOUR, CardShape.CLOVER),
-                    ).toGamerCards()
-                } else {
-                    listOf(
-                        Card(CardNumber.Q, CardShape.HEART),
-                        Card(CardNumber.THREE, CardShape.HEART),
-                    ).toGamerCards()
-                }
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.FIVE, CardShape.CLOVER),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 14
-        participants.last().getCurrentCards().getCurrentScore() shouldBe 13
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
-
-    "참가자가 블랙잭 인경우 1.5배를 잃는다" {
-        val input = "pita"
-        val expected = -1500
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                listOf(
-                    Card(CardNumber.Q, CardShape.CLOVER),
-                    Card(CardNumber.A, CardShape.CLOVER),
-                ).toGamerCards()
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.FIVE, CardShape.CLOVER),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 21
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
-
-    "참가자와 딜러가 블랙잭 인경우 둘다 0원이다" {
-        val input = "pita"
-        val expected = 0
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                listOf(
-                    Card(CardNumber.Q, CardShape.CLOVER),
-                    Card(CardNumber.A, CardShape.CLOVER),
-                ).toGamerCards()
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.A, CardShape.HEART),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 21
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
-
-    "참가자가 HIT, BUST 일 때, 딜러가 BUST인 경우 HIT만 돈을 지급한다" {
-        val input = "pita,haero"
-        val expected = 1000
-        var firstInput = true
-
-        val allParticipantWithBetAmount = AllParticipantWithBetAmount.newInstance(
-            names = input,
-            betAmounts = listOf(ParticipantBetAmount(1000), ParticipantBetAmount(2000))
-        )
-        val participants = Participants.newInstance(
-            allParticipantWithBetAmount,
-            cardDeque = {
-                if (firstInput) {
-                    firstInput = false
-                    listOf(
-                        Card(CardNumber.Q, CardShape.CLOVER),
-                        Card(CardNumber.SIX, CardShape.CLOVER),
-                    ).toGamerCards()
-                } else {
-                    listOf(
-                        Card(CardNumber.Q, CardShape.HEART),
-                        Card(CardNumber.J, CardShape.HEART),
-                        Card(CardNumber.THREE, CardShape.HEART),
-                    ).toGamerCards()
-                }
-            }
-        )
-        val dealer = Dealer.newInstance(
-            listOf(
-                Card(CardNumber.K, CardShape.CLOVER),
-                Card(CardNumber.FIVE, CardShape.CLOVER),
-                Card(CardNumber.TEN, CardShape.CLOVER),
-            ).toGamerCards()
-        )
-        participants.first().getCurrentCards().getCurrentScore() shouldBe 16
-        participants.last().getCurrentCards().getCurrentScore() shouldBe 23
-        participants.getParticipantsLoss(dealer) shouldBe expected
-    }
 })
